+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## develop branch
 
-Nothing yet.
+### Backwards-Compatibility Breaks
+
+* Objects namespace is now ClassesAndObjects
+
+### New
+
+* Classes\Objects\InvokeMethod::onString()
 
 ## 1.0.0 - Sat Jun 27 2015